spin_runtime_config

Struct ResolvedRuntimeConfig

pub struct ResolvedRuntimeConfig<T> {
    pub runtime_config: T,
    pub key_value_resolver: RuntimeConfigResolver,
    pub sqlite_resolver: RuntimeConfigResolver,
    pub state_dir: Option<PathBuf>,
    pub log_dir: Option<PathBuf>,
    pub max_instance_memory: Option<usize>,
    pub toml: Table,
}

A runtime configuration which has been resolved from a runtime config source.

Includes other pieces of configuration that are used to resolve the runtime configuration.

Fields
runtime_config: T

The resolved runtime configuration.

key_value_resolver: RuntimeConfigResolver

The resolver used to resolve key-value stores from runtime configuration.

sqlite_resolver: RuntimeConfigResolver

The resolver used to resolve sqlite databases from runtime configuration.

state_dir: Option<PathBuf>

The fully resolved state directory.

None is used for an “unset” state directory which each factor will treat differently.

log_dir: Option<PathBuf>

The fully resolved log directory.

None is used for an “unset” log directory.

max_instance_memory: Option<usize>

The maximum memory allocation limit.

toml: Table

The input TOML, for informational summaries.

Implementations

impl<T> ResolvedRuntimeConfig<T>

pub fn summarize(&self, runtime_config_path: Option<&Path>)

impl<T> ResolvedRuntimeConfig<T>
where T: for<'a, 'b> TryFrom<TomlRuntimeConfigSource<'a, 'b>>, for<'a, 'b> <T as TryFrom<TomlRuntimeConfigSource<'a, 'b>>>::Error: Into<Error>,

pub fn from_file( runtime_config_path: Option<&Path>, local_app_dir: Option<PathBuf>, provided_state_dir: UserProvidedPath, provided_log_dir: UserProvidedPath, ) -> Result<Self>

Creates a new resolved runtime configuration from a runtime config source TOML file.

provided_state_dir is the explicitly provided state directory, if any.

pub fn new( toml_resolver: TomlResolver<'_>, runtime_config_path: Option<&Path>, ) -> Result<Self>

Creates a new resolved runtime configuration from a TOML table.

pub fn state_dir(&self) -> Option<PathBuf>

The fully resolved state directory.

pub fn log_dir(&self) -> Option<PathBuf>

The fully resolved state directory.

pub fn max_instance_memory(&self) -> Option<usize>

The maximum memory allocation limit.